+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - **`[emoji]`** Code refactoring
 - **`[env]`** Code refactoring
 - **`[errors]`** Code refactoring
+- **`[events]`** Code refactoring
 - **`[fmtc]`** Code refactoring
 - **`[log]`** Code refactoring
 - **`[req]`** Code refactoring